local: Make rw_timeout_ms unsigned int
The set_timeout() API takes a unsigned int so using a unsigned int for
storing the value is large enough, no need to use a unsigned long.

+++ local.c
@@ -78,7 +78,7 @@ struct block {
 };
 
 struct iio_context_pdata {
-	unsigned long rw_timeout_ms;
+	unsigned int rw_timeout_ms;
 };
 
 struct iio_device_pdata {
